fix: refactor folders naming convention
