--- frontend/src/File Manager/Actions/InlineEdit.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const InlineEdit = () => {
-  return <div>InlineEdit</div>;
-};
-
-export default InlineEdit;
