+++ frontend/src/FileManager/Actions/CreateFolder/CreateFolder.action.jsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from "react";
-import { useDetectOutsideClick } from "../../hooks/useDetectOutsideClick";
-import { duplicateNameHandler } from "../../utils/duplicateNameHandler";
-import NameInput from "../../components/Name Input/NameInput";
-import ErrorTooltip from "../../components/Error Tooltip/ErrorTooltip";
-import { useFileNavigation } from "../../contexts/FileNavigationContext";
-import { useLayout } from "../../contexts/LayoutContext";
+import { useDetectOutsideClick } from "../../../hooks/useDetectOutsideClick";
+import { duplicateNameHandler } from "../../../utils/duplicateNameHandler";
+import NameInput from "../../../components/NameInput/NameInput";
+import ErrorTooltip from "../../../components/ErrorTooltip/ErrorTooltip";
+import { useFileNavigation } from "../../../contexts/FileNavigationContext";
+import { useLayout } from "../../../contexts/LayoutContext";
 
 const maxNameLength = 220;
