+++ frontend/src/FileManager/Actions/Delete/Delete.action.jsx
