+++ frontend/src/FileManager/Actions/Delete/Delete.action.scss
