+++ frontend/src/FileManager/Actions/PreviewFile/PreviewFile.action.scss
