+++ frontend/src/FileManager/BreadCrumb/BreadCrumb.scss
