+++ frontend/src/FileManager/FileList/FileItem.jsx
@@ -2,13 +2,13 @@ import { useEffect, useState } from "react";
 import { FaRegFile, FaRegFolderOpen, FaRegPaste } from "react-icons/fa6";
 import { PiFolderOpen } from "react-icons/pi";
 import { MdOutlineDelete } from "react-icons/md";
-import ContextMenu from "../../components/Context Menu/ContextMenu";
+import ContextMenu from "../../components/ContextMenu/ContextMenu";
 import { useDetectOutsideClick } from "../../hooks/useDetectOutsideClick";
 import { BiRename } from "react-icons/bi";
 import { BsCopy, BsScissors } from "react-icons/bs";
 import { useFileIcons } from "../../hooks/useFileIcons";
-import CreateFolderAction from "../Actions/CreateFolder.action";
-import RenameAction from "../Actions/Rename.action";
+import CreateFolderAction from "../Actions/CreateFolder/CreateFolder.action";
+import RenameAction from "../Actions/Rename/Rename.action";
 import { getDataSize } from "../../utils/getDataSize";
 import { formatDate } from "../../utils/formatDate";
 import { useFileNavigation } from "../../contexts/FileNavigationContext";
