+++ frontend/src/FileManager/FileList/FileList.scss
