+++ frontend/src/FileManager/FileManager.scss
