+++ frontend/src/FileManager/NavigationPane/NavigationPane.scss
