+++ frontend/src/FileManager/Toolbar/Toolbar.scss
