+++ frontend/src/components/ErrorTooltip/ErrorTooltip.scss
