+++ frontend/src/components/NameInput/NameInput.scss
